% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas_agggregate.r
\name{atlas_aggregate}
\alias{atlas_aggregate}
\title{atlas_aggregate}
\usage{
atlas_aggregate(
  x,
  admin_level = c("admin0", "admin1", "admin2"),
  fun,
  colname = NULL,
  ...
)
}
\arguments{
\item{x}{A SpatRaster or path to raster layer}

\item{admin_level}{A character string indicating the administrative level to aggregate to}

\item{fun}{The aggregation function used by exact_extract}

\item{colname}{The column name in the output dataframe. If missing, it is inherited from the raster}

\item{...}{Additional arguments passed to exact_extract}
}
\value{
An sf object containing the aggregated data
}
\description{
This function normalizes the input vector or SpatRaster object using the min-max normalization technique.
If the input is a SpatRaster, the function calculates the minimum and maximum values and performs normalization.
If the input is a numeric vector, the function calculates the minimum and maximum values and performs normalization.
}
\seealso{
\code{\link[exactextractr]{exact_extract}}
}
