% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_builder.r
\name{atlas_metadata}
\alias{atlas_metadata}
\title{Generate meta-data for Adaptation Atlas Layers}
\usage{
atlas_metadata(
  data = NA,
  folder = NA,
  save_output = c("csv", "txt"),
  dataset.title_short = "",
  dataset.title_long = "",
  dataset.desc = "",
  dataset.author = "",
  dataset.contact = "",
  dataset.contact_email = "",
  dataset.pub_doi = "",
  dataset.data_doi = "",
  dataset.sourceurl = "",
  dataset.projecturl = "",
  dataset.citation = "",
  dataset.licence = "",
  file.filename = "",
  file.format = "",
  file.data_type = "",
  file.no_value_data = "",
  file.file_naming_convention = "",
  file.flags = "",
  variable.theme = "",
  variable.subtheme = "",
  variable.name = "",
  variable.subname = "",
  variable.commodity = "",
  variable.type = "",
  variable.statistic = "",
  variable.unit = "",
  method.analysis_type = "",
  method.description = "",
  method.github = "",
  method.qual_indicator = "",
  method.qual_availability = "",
  grid.proj = "",
  grid.epsg = "",
  grid.xres = "",
  grid.yres = "",
  grid.xmin = "",
  grid.xmax = "",
  grid.ymin = "",
  grid.ymax = "",
  grid.nrow = "",
  grid.ncol = "",
  temporal.resolution = "",
  temporal.start_date = "",
  temporal.end_date = "",
  data.categorical_val = "",
  data.categorical_desc = "",
  data.shapefile_field = "",
  data.shapefile_type = "",
  data.shapefile_unit = "",
  data.shapefile_description = ""
)
}
\arguments{
\item{data}{A \code{SpatRaster} or \code{SparVector} object.}

\item{folder}{The directory path for where the output metadata should be saved.}

\item{save_output}{If folder is provided, should the metadata be saved as a csv or txt file?}

\item{dataset.title_short}{A short title to label the dataset. e.g., \code{MapSPAM}.}

\item{dataset.title_long}{A full title to label the dataset. e.g., \verb{Travel time to cities}.}

\item{dataset.desc}{A short description enabling the user to quickly understand the subject of the dataset and how to interpret it. For example, \verb{The value of each pixel is the estimated travel time in minutes to the nearest urban area in 2015. There are 12 data layers based on different sets of urban areas, defined by their population in year 2015}.}

\item{dataset.author}{Name of authors. Surname of lead author only.}

\item{dataset.contact}{If the dataset has a gate-keeper or contact author list their full name here.}

\item{dataset.contact_email}{If the dataset has a gate-keeper or main author list their email address here.}

\item{dataset.pub_doi}{The doi for the publication(s) associated with this datasest. Use a list to provide multiple entries. For example, \code{list("10.1038/nature25181","10.1038/EE23432")}}

\item{dataset.data_doi}{The doi for this dataset. This could be a Harvard dataverse doi, for example, \code{10.7910/DVN/FSSKBW}.}

\item{dataset.sourceurl}{The link that gets the user closest to the dataset if this is not the dataset doi. For example, a page describing a dataset API or data portal such as \verb{https://www.fao.org/faostat/en/#data/QCL}. If there is a github associated with analysis please list this in the methods section.}

\item{dataset.projecturl}{If the dataset has a project website then record the url here, for example \verb{https://www.mapspam.info/}.}

\item{dataset.citation}{Main citation(s) for this dataset in Havard format. Use a list to separate multiple entries.}

\item{dataset.licence}{Is the dataset open-access? The information to is to include the licence type, other wise \code{open} or \code{private} will suffice.}

\item{file.filename}{The filename of the dataset \verb{e.g., aez_global.shp}.}

\item{file.format}{The format of the file, for example \verb{LZW compressed GeoTIFF, shapefile, netcdf4}.}

\item{file.data_type}{Type and/or size of the dataset, for example  \verb{Byte (16 bit Unsigned Integer)}.}

\item{file.no_value_data}{The value that indicates no-data or NA values in the dataset, for example \code{65535} or \code{9999}.}

\item{file.file_naming_convention}{A free text description of any file naming convention that applies, so that a user can easily interpret the filename.}

\item{file.flags}{Additional flags for the file which controls how the file is opened.}

\item{variable.theme}{One of \code{hazard}, \code{exposure}, \code{adaptative_capacity}, \code{solutions}, \verb{enabling environment} or combinations of these. Use a \code{.} delimiter to indicate a combination for example, \code{hazard.exposure}.}

\item{variable.subtheme}{Please use as few words as possible to describe a subtheme, for example \code{rainfall}, \verb{heat stress}, or \verb{access to markets}.}

\item{variable.name}{The specific name of the variable, for example \verb{crop yield}, \verb{biophysical suitability} or \verb{aridity index}. If the dataset refers to a specific crop the crop name should be recorded in the \code{variable.commodity} parameter.}

\item{variable.subname}{If more groupings are needed to describe the data structure then this column can be used. For example MapSPAM value of production data could have  \verb{variable.name=value of production} and  \verb{variable.subname = all technologies} or \verb{variable.subname = rainfed only}. Do not record information about statistics here, use the \code{variable.statistic} parameter instead.}

\item{variable.commodity}{If the variable relates to a specific crop please include the crop name here (scientific name preferred). For example, \verb{Zea mays}}

\item{variable.type}{The type of variable, usually one of \code{continuous}, \code{categorical}, or \code{ordinal}. For ordinal or categorical rasters please complete the \code{data.categorical_val} and \code{data.categorical_desc} parameters.}

\item{variable.statistic}{If a variable has different files that represent different calculations or groupings, such as the mean and confidence intervals of a variable, then please note the calculation here. For example, \code{mean}, \code{median}, \verb{95\% CI high},\verb{95\% CI low}, etc.}

\item{variable.unit}{Unit of measurement, for example \code{kg/ha}, \code{kg/kg}, \code{days}, \code{minutes}, etc. See \verb{https://www.ebi.ac.uk/ols/ontologies/om} for standardized SI units. Please use a forward slash \code{/} delim rather than \code{-1} (so \code{kg/ha} not \verb{kg ha-1}).}

\item{method.analysis_type}{A general description of the type of analysis employed. For example, \verb{niche model} or\verb{DSSAT crop model}.}

\item{method.description}{A summary of methods used to create the dataset.}

\item{method.github}{If there is a github repo associated with the analysis please list it here. Ideally it should walk a user though the steps to recreate a datasest.}

\item{method.qual_indicator}{The name of a data avaiable indicator, typically for niche or analogue layers derived from empirical data. For example \verb{number of studies}.}

\item{method.qual_availability}{The value for the data availability indicator defined in the row above. This can be numeric or ordinal (e.g. \code{high}, \code{medium}, or \code{low})}

\item{grid.proj}{This field is auto-filled if the dataset is provided as the \code{data} parameter. The proj4 string of the dataset.}

\item{grid.epsg}{This field is auto-filled if the dataset is provided as the \code{data} parameter. The EPSG code of the dataset.}

\item{grid.xres}{This field is auto-filled if the dataset is provided as the \code{data} parameter. For gridded datasets, the spatial resolution of celll latitude dataset in decimal degrees.}

\item{grid.yres}{This field is auto-filled if the dataset is provided as the \code{data} parameter. For gridded datasets, the spatial resolution of celll longitude dataset in decimal degrees.}

\item{grid.xmin}{This field is auto-filled if the dataset is provided as the \code{data} parameter. Minimum latitude of dataset in decimal degrees.}

\item{grid.xmax}{This field is auto-filled if the dataset is provided as the \code{data} parameter. Maximum latitude of dataset in decimal degrees.}

\item{grid.ymin}{This field is auto-filled if the dataset is provided as the \code{data} parameter. Minimum longtiude of dataset in decimal degrees.}

\item{grid.ymax}{This field is auto-filled if the dataset is provided as the \code{data} parameter. Maximum longtiude of dataset in decimal degrees.}

\item{grid.nrow}{This field is auto-filled if the dataset is provided as the \code{data} parameter. Number of rows in gridded data.}

\item{grid.ncol}{This field is auto-filled if the dataset is provided as the \code{data} parameter. Number of cols in gridded data.}

\item{temporal.resolution}{What time period does the dataset refer to? For example \code{annual}, \code{dekadal}, \code{daily}, \verb{long-term average} or similar.}

\item{temporal.start_date}{The start date of the temporal period for which the dataset is revelant to or derived from. Preferably a date as \code{yyyy--mm-dd}, \code{yyyy-mm} or \code{yyyy}. For example, \code{2018-01-01}, \code{2018-01} or \code{2018}. Please use r date format \code{yyyy-mm-dd}.}

\item{temporal.end_date}{The end date of the temporal period for which the dataset is revelant to or derived from. Preferably a date as \code{yyyy--mm-dd}, \code{yyyy-mm} or \code{yyyy}. For example, \code{2018-01-01}, \code{2018-01} or \code{2018}. Please use r date format \code{yyyy-mm-dd}.}

\item{data.categorical_val}{For gridded categorical or ordinal datasets provide a list of values that correspond to values in the dataset, e.g. \code{list(1,2,3,4,5)}.}

\item{data.categorical_desc}{For the values in \code{data.categorical_val} provide a list of descriptions that correspond to these values, e.g. \code{list("none","mild","moderate","severe","extreme")}.}

\item{data.shapefile_field}{For the data.table associated with a shapefile please list the names of key fields, e.g. \code{list("Admin0","Admin1","\% of population")}.}

\item{data.shapefile_type}{For the fields listed in \code{data.shapefile_field} please provide a list of data types, e.g. \code{list("categorical","categorical","continuous")}.}

\item{data.shapefile_unit}{For the fields listed in \code{data.shapefile_field} please provide a list of units, e.g. \code{list(NA,NA,"proportion")}.}

\item{data.shapefile_description}{For the fields listed in \code{data.shapefile_field} please provide a description of the field, e.g. \code{list("Country name","Administrative unit 1 name","Poverty headcount ratio at national poverty lines")}.}
}
\value{
A data.frame of meta.data
imports sf st_crs st_bbox
imports stars st_res
imports terra crs ext nrow ncol res
imports tools file_path_sans_ext file_ext
imports utils write.csv
}
\description{
This function will assist you to systematically generate meta-data for the Adaptation Atlas.
}
\details{
The \code{data} parameter can be used to supply a Spatraster, SpatVector, Stars, or sf object from which geospatial dataset will automatically extracted.
The \code{data} parameter can also be a file path to a spatial file such as .gpkg, .tif, .nc4, etc.
In all cases provide multiple values as a list, for example \code{list(10.1038/nature25181, 10.1038/EE23432)}
}
