% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas_resample.r
\name{atlas_resample}
\alias{atlas_resample}
\title{Resample Raster Data to Adaptation Atlas Grid (5 arc min)}
\usage{
atlas_resample(x, fun = "bilinear", ...)
}
\arguments{
\item{x}{An input raster object or a data structure that can be
converted to a spatial raster (e.g., a matrix or data frame).}

\item{fun}{The resampling method to use. Default is "bilinear," but other methods
supported by the 'terra' package (e.g., "sum", "bilinear" "cubic", "mean") can be specified.}

\item{...}{Additional arguments to be passed to the terra::resample function.}
}
\value{
A resampled raster object on the standard atlas grid.
}
\description{
This function resamples a given raster object to a standard atlas grid with a specified resampling method.
}
\examples{
# Example 1: Resample a raster using the default method
 \dontrun{result <- atlas_resample(input_raster)}
#
# Example 2: Resample a raster using a specified resampling method
 \dontrun{result <- atlas_resample(input_raster, fun = "sum")}

}
\seealso{
\code{\link[terra:resample]{terra::resample()}} for detailed information on the resampling process
}
