% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_s3.r
\name{explore_s3}
\alias{explore_s3}
\title{Explore S3 Bucket}
\usage{
explore_s3(
  bucket.path = NULL,
  public = FALSE,
  region = "us-east-1",
  recursive = TRUE,
  open.view = TRUE,
  return_type = c("list", "json", "data.frame", "tree")
)
}
\arguments{
\item{bucket.path}{The path of the S3 bucket to explore. If not provided, a default path to the "Adaptation Atlas" data will be used}

\item{public}{Logical value indicating whether the bucket is public or not. If the bucket it public, it prevents the use of AWS credentials avoiding an error.}

\item{region}{The AWS region where the bucket is located. Default is "us-east-1".}

\item{recursive}{Logical value indicating whether to explore the bucket recursively or not. Default is TRUE.}

\item{open.view}{Logical value indicating whether to open the list in a new window or not. Default is TRUE.}

\item{return_type}{The format in which the information should be returned. Possible values are "list", "json", "data.frame", "tree". Default is "list".}
}
\value{
Depending on the specified return type, the function returns a list, hierarchical JSON, data frame, or tree object.
}
\description{
This function is used to explore the contents of a S3 bucket. It is designed to return an object or viewer that
is as clos to a normal directory tree as possible. It can also be used to return a dataframe that contiains the
data and structure of the bucket.
}
\details{
opens a list of AWS S3 objects in a new window that is in the format of a collapsible directory tree shape
}
\examples{
explore_s3(bucket.path = "noaa-gsod-pds", public = TRUE, recursive = FALSE, open.view = FALSE)

\dontrun{
 Sys.setenv(
    AWS_ACCESS_KEY_ID = "your AWS access key",
    AWS_SECRET_ACCESS_KEY = "your AWS secret key",
    AWS_DEFAULT_REGION = "your AWS region" (optional),
    OR
    AWS_CONFIG_FILE = "~/.aws/config" (or ~/path/to/aws/config/file),
 )
explore_s3() # this will default to the Adaptation Atlas Data
}
}
