% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_to_s3.r
\name{file_to_s3}
\alias{file_to_s3}
\title{Upload a file to an S3 bucket}
\usage{
file_to_s3(localpath, bucketpath, cog = TRUE, ...)
}
\arguments{
\item{localpath}{A character string specifying the path to the local file}

\item{bucketpath}{A character string specifying the path to the destination bucket in S3}

\item{cog}{A logical value indicating whether to convert a tif to a COG before uploading (default is TRUE).}

\item{...}{Additional arguments passed to the \verb{s3fs::s3filesystem_class file_upload} function}
}
\value{
An object representing the uploaded file in S3
}
\description{
This function uploads a file to the specified S3 bucket. If the file extension is "tif" and the \code{cog} parameter is set to TRUE, the function converts the file to a COG (Cloud-Optimized GeoTIFF) before uploading.
}
\examples{
\dontrun{
# setting credentials is only required if there is no config file saved to "~/.aws/config"
 Sys.setenv(
    AWS_ACCESS_KEY_ID = "your AWS access key",
    AWS_SECRET_ACCESS_KEY = "your AWS secret key",
    AWS_DEFAULT_REGION = "your AWS region" (optional),
    OR
    AWS_CONFIG_FILE = "~/.aws/config" (or ~/path/to/aws/config/file),
 )
file_to_s3(localpath = "path/to/local/file.tif", bucketpath = "path/to/bucket/file.tif", cog = TRUE, overwrite = FALSE)

# or for a full folder upload
folder_list <- list.files("path/to/folder", recursive = TRUE, pattern = "\\\\.tif|\\\\.gpkg")
for (i in seq_along(folder_list)) {
  file_to_s3(localpath = folder_list[i],
  bucketpath = paste0("path/to/bucket/", folder_list[i]),
  cog = TRUE, overwrite = FALSE)
}
}
}
