% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax_norm.r
\name{MinMax_norm}
\alias{MinMax_norm}
\title{MinMax_norm}
\usage{
MinMax_norm(x, inverse = FALSE)
}
\arguments{
\item{x}{A numeric vector or SpatRaster to be normalized.}

\item{inverse}{Logical value indicating whether to apply inverse the values}
}
\value{
A normalized numeric vector or SpatRaster object.
}
\description{
This function normalizes the input vector or SpatRaster object using the min-max normalization technique.
If the input is a SpatRaster, the function calculates the minimum and maximum values and performs normalization.
If the input is a numeric vector, the function calculates the minimum and maximum values and performs normalization.
}
