% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bivariate_maps.r
\name{bivar_legend}
\alias{bivar_legend}
\title{Create a bivariate legend plot.}
\usage{
bivar_legend(x.title, y.title, n = 3, pal = NULL, font_size = 12)
}
\arguments{
\item{x.title}{A character string specifying the title for the x-axis.}

\item{y.title}{A character string specifying the title for the y-axis.}

\item{n}{The number of bivariate groups (default is 3). Must be between
3 and 9. A value of 3 is recommended for readability.}

\item{pal}{A vector of color values for the legend. If NULL, a default color
palette will be used for n = 3, but you must provide a palette for n other than 3.}

\item{font_size}{The font size for axis titles (default is 12).}
}
\value{
A ggplot2 object representing the bivariate legend.
}
\description{
This function generates a legend plot to represent bivariate data. It is a
grid of colored squares that can be used to represent the values of two
categorical variables in a bivariate plot. It is designed to be paired with
the \code{make_bivariate_data} function.
}
\examples{
# Generate a bivariate legend with default parameters
bivar_legend("X Variable", "Y Variable")

# Create a full bivariate map with legend using sf object
\dontrun{
library(patchwork)
legend <- bivar_legend("X Variable", "Y Variable")
bivar_data <- make_bivariate_data()
map <- ggplot(bivar_data) +
 geom_sf(aes(fill = bivar_colour), color = NA) +
 scale_fill_identity(guide = "legend") +
 theme_void() +
 theme(legend.position = "none")

# create a patchwork layout
layout <- "
#22
122
"
# Use patchwork to combine the map and legend
legend + map + plot_layout(design = layout, widths = c(.5, 1))
}

# Create a legend using n = 4
\dontrun{
pal <- c("#e8e8e8", "#bddede", "#8ed4d4", "#5ac8c8", "#dabdd4",
  "#bdbdd4", "#8ebdd4", "#5abdc8", "#cc92c1", "#bd92c1", "#8e92c1",
  "#5a92c1", "#be64ac","#bd64ac", "#8e64ac", "#5a64ac")
 bivar_legend("Human Heat Stress", "Minutes to Healthcare",
   n = 4, pal = pal)
}

}
\seealso{
\code{\link[=make_bivariate_data]{make_bivariate_data()}} to create bivariate data from raster or vector data
}
