% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_bivar_data.r
\name{make_bivariate_data}
\alias{make_bivariate_data}
\title{Create a bivariate data}
\usage{
make_bivariate_data(data, n = 3, x.val = NULL, y.val = NULL, pal = NULL)
}
\arguments{
\item{data}{A 2 layer SpatRaster or a SpatVector or sf object.}

\item{n}{Number of classes for the bivariate map. Suggested is 3.}

\item{x.val}{First attribute name (vector data only).}

\item{y.val}{Second attribute name (vector data only).}

\item{pal}{A vector of color codes to use for the map. If not provided, a
default color palette is used if n = 3.}
}
\value{
A raster or vector object containing bivariate categories. If input
is a raster, the output as a categorical raster where the levels are
the bivariate groups. If input is a vector, the output vector contains the
initial attributes plus an attribute of "bi_color" with the bivariate
colors, "bi_groups" with the combined x and y categories, and 2 "*_cats"
with the x and y categories.
}
\description{
This function creates an raster or vector with bivariate categories
based on an x and y value. Paired with the bivar_legend function, this
function can be used alongside ggplot to create a full bivariate map as shown
in the below example.
}
\examples{
# Create a full bivariate map with legend using sf object
\dontrun{
# map_data <- make_bivariate_data(t_sf, x.val = "heat_stress", y.val = "")
# legend <- bivar_legend(x.title = "Heat Stress", y.title = "human_index")
# map <- ggplot(map_data) +
#   geom_sf(aes(fill = bi_color), color = NA) +
#   scale_fill_identity(guide = "legend") +
#   theme_void() +
#   theme(legend.position = "none")
# layout <- "
# #22
# 122
# "
# legend + map + plot_layout(design = layout, widths = c(.5, 1))
}

}
\seealso{
\code{\link[=bivar_legend]{bivar_legend()}} to create bivariate legend
}
