% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cog.r
\name{write_cog}
\alias{write_cog}
\title{Write a Cloud-Optimized GeoTIFF (COG) from a source raster file.}
\usage{
write_cog(path, out_dir = NULL, add_flag = TRUE)
}
\arguments{
\item{path}{A character string specifying the path to the source raster file.}

\item{out_dir}{A character string specifying the output directory where the COG
will be saved. If NULL (default), the COG will be saved in the same directory
as the source file.}

\item{add_flag}{A logical value indicating whether to add "_cog" to file's
name. Default is TRUE.}
}
\value{
NULL. The function generates a COG file and saves it to the specified
location.
}
\description{
This function takes a source raster file, generates overviews for it, and
writes a Cloud-Optimized GeoTIFF (COG) to the specified output directory.
}
\examples{
# Create a COG from a source raster file with default parameters
\dontrun{write_cog("path/to/source_raster.tif")}
#
# Create a COG with a custom output directory and no flag added to the filename
\dontrun{write_cog("path/to/source_raster.tif", out_dir = "cogs", add_flag = FALSE)}
#
}
